+++ com.unity.netcode.gameobjects/Runtime/Core/NetworkObject.cs
@@ -54,8 +54,6 @@ namespace Unity.Netcode
         /// </summary>
         internal NetworkManager NetworkManagerOwner;
 
-        private ulong m_NetworkObjectId;
-
         /// <summary>
         /// Gets the unique Id of this object that is synced across the network
         /// </summary>
@@ -288,7 +286,6 @@ namespace Unity.Netcode
                 throw new VisibilityChangeException("Cannot hide an object from the server");
             }
 
-
             Observers.Remove(clientId);
 
             if (NetworkManager.NetworkConfig.UseSnapshotSpawn)
@@ -319,7 +316,7 @@ namespace Unity.Netcode
                 throw new ArgumentNullException("At least one " + nameof(NetworkObject) + " has to be provided");
             }
 
-            NetworkManager networkManager = networkObjects[0].NetworkManager;
+            var networkManager = networkObjects[0].NetworkManager;
 
             if (!networkManager.IsServer)
             {
@@ -364,7 +361,8 @@ namespace Unity.Netcode
                 throw new NotServerException($"Destroy a spawned {nameof(NetworkObject)} on a non-host client is not valid. Call {nameof(Destroy)} or {nameof(Despawn)} on the server/host instead.");
             }
 
-            if (NetworkManager != null && NetworkManager.SpawnManager != null && NetworkManager.SpawnManager.SpawnedObjects.TryGetValue(NetworkObjectId, out var networkObject))
+            if (NetworkManager != null && NetworkManager.SpawnManager != null &&
+                NetworkManager.SpawnManager.SpawnedObjects.TryGetValue(NetworkObjectId, out var networkObject))
             {
                 NetworkManager.SpawnManager.OnDespawnObject(networkObject, false);
             }
@@ -372,19 +370,18 @@ namespace Unity.Netcode
 
         private SnapshotDespawnCommand GetDespawnCommand()
         {
-            var command = new SnapshotDespawnCommand();
-            command.NetworkObjectId = NetworkObjectId;
-
-            return command;
+            return new SnapshotDespawnCommand { NetworkObjectId = NetworkObjectId };
         }
 
         private SnapshotSpawnCommand GetSpawnCommand()
         {
-            var command = new SnapshotSpawnCommand();
-            command.NetworkObjectId = NetworkObjectId;
-            command.OwnerClientId = OwnerClientId;
-            command.IsPlayerObject = IsPlayerObject;
-            command.IsSceneObject = (IsSceneObject == null) || IsSceneObject.Value;
+            var command = new SnapshotSpawnCommand
+            {
+                NetworkObjectId = NetworkObjectId,
+                OwnerClientId = OwnerClientId,
+                IsPlayerObject = IsPlayerObject,
+                IsSceneObject = (IsSceneObject == null) || IsSceneObject.Value
+            };
 
             ulong? parent = NetworkManager.SpawnManager.GetSpawnParentId(this);
             if (parent != null)
@@ -415,8 +412,7 @@ namespace Unity.Netcode
         private void SnapshotSpawn(ulong clientId)
         {
             var command = GetSpawnCommand();
-            command.TargetClientIds = new List<ulong>();
-            command.TargetClientIds.Add(clientId);
+            command.TargetClientIds = new List<ulong> { clientId };
             NetworkManager.SnapshotSystem.Spawn(command);
         }
 
@@ -429,8 +425,7 @@ namespace Unity.Netcode
         internal void SnapshotDespawn(ulong clientId)
         {
             var command = GetDespawnCommand();
-            command.TargetClientIds = new List<ulong>();
-            command.TargetClientIds.Add(clientId);
+            command.TargetClientIds = new List<ulong> { clientId };
             NetworkManager.SnapshotSystem.Despawn(command);
         }
 
@@ -821,13 +816,13 @@ namespace Unity.Netcode
             }
         }
 
-        internal void WriteNetworkVariableData(FastBufferWriter writer, ulong clientId)
+        internal void WriteNetworkVariableData(FastBufferWriter writer, ulong targetClientId)
         {
             for (int i = 0; i < ChildNetworkBehaviours.Count; i++)
             {
                 var behavior = ChildNetworkBehaviours[i];
                 behavior.InitializeVariables();
-                behavior.WriteNetworkVariableData(writer, clientId);
+                behavior.WriteNetworkVariableData(writer, targetClientId);
             }
         }